done except null case in ptr

+++ ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: enoshahi <enoshahi@student.42.fr>          +#+  +:+       +#+        */
+/*   By: enoshahi < enoshahi@student.42abudhabi.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/18 14:02:45 by enoshahi          #+#    #+#             */
-/*   Updated: 2024/09/03 17:41:47 by enoshahi         ###   ########.fr       */
+/*   Updated: 2024/09/04 17:22:44 by enoshahi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int	ft_printf(const char *str, ...);
 int	ft_putchar(char c);
 int	ft_puthex(unsigned long num, char base);
 int	ft_putnbr(int n);
-int	ft_putptr(unsigned int ptr);
+int	ft_putptr(unsigned long ptr);
 int	ft_putstr(char *s);
 int	ft_putusd(unsigned int n);
